# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Antenna Array Optimizer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Number of antenna elements per ring (5 rings, 2-8 elements each)
 */
export interface AntennaConfig {
  /**
   * @minimum 2
   * @maximum 8
   */
  ring1: number;
  /**
   * @minimum 2
   * @maximum 8
   */
  ring2: number;
  /**
   * @minimum 2
   * @maximum 8
   */
  ring3: number;
  /**
   * @minimum 2
   * @maximum 8
   */
  ring4: number;
  /**
   * @minimum 2
   * @maximum 8
   */
  ring5: number;
}

export interface PerformanceMetrics {
  /** Half-power beamwidth in degrees */
  hpbw: number;
  /** Main lobe gain in dB */
  mainLobeGain: number;
  /** Side lobe level in dB */
  sideLobeLevel: number;
}

export interface RadiationPattern {
  /** Azimuth angle values in degrees (0-360) */
  thetaDeg: number[];
  /** Array factor in dB (normalized) */
  afDb: number[];
}

export interface SimulateAntennaBody {
  config: AntennaConfig;
  /** Steering angle in degrees */
  theta0Deg?: number;
}

export interface SimulationResult {
  config: AntennaConfig;
  theta0Deg: number;
  pattern: RadiationPattern;
  metrics: PerformanceMetrics;
}

export interface CompareConfigurationsBody {
  reference: AntennaConfig;
  predicted: AntennaConfig;
  /** Steering angle in degrees */
  theta0Deg?: number;
}

export type ComparisonResultReference = {
  config: AntennaConfig;
  pattern: RadiationPattern;
  metrics: PerformanceMetrics;
};

export type ComparisonResultPredicted = {
  config: AntennaConfig;
  pattern: RadiationPattern;
  metrics: PerformanceMetrics;
};

export type ComparisonResultErrors = {
  hpbwError: number;
  mainLobeGainError: number;
  sidelobeLevelError: number;
  globalError: number;
};

export interface ComparisonResult {
  theta0Deg: number;
  reference: ComparisonResultReference;
  predicted: ComparisonResultPredicted;
  errors: ComparisonResultErrors;
}

export interface Prediction {
  id: number;
  createdAt: string;
  theta0Deg: number;
  referenceConfig: AntennaConfig;
  predictedConfig: AntennaConfig;
  referenceMetrics: PerformanceMetrics;
  predictedMetrics: PerformanceMetrics;
  globalError: number;
}

export interface PredictionStats {
  totalPredictions: number;
  avgGlobalError: number;
  avgHpbwError: number;
  avgMainLobeGainError: number;
  avgSidelobeLevelError: number;
  /** @nullable */
  bestPredictionId: number | null;
  /** @nullable */
  worstPredictionId: number | null;
}

export type GetRecentHistoryParams = {
  limit?: number;
};