# lib/api-zod/src/generated/types/comparisonResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Antenna Array Optimizer API
 * OpenAPI spec version: 0.1.0
 */
import type { ComparisonResultErrors } from "./comparisonResultErrors";
import type { ComparisonResultPredicted } from "./comparisonResultPredicted";
import type { ComparisonResultReference } from "./comparisonResultReference";

export interface ComparisonResult {
  theta0Deg: number;
  reference: ComparisonResultReference;
  predicted: ComparisonResultPredicted;
  errors: ComparisonResultErrors;
}