# lib/api-zod/src/generated/types/predictionStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Antenna Array Optimizer API
 * OpenAPI spec version: 0.1.0
 */

export interface PredictionStats {
  totalPredictions: number;
  avgGlobalError: number;
  avgHpbwError: number;
  avgMainLobeGainError: number;
  avgSidelobeLevelError: number;
  /** @nullable */
  bestPredictionId: number | null;
  /** @nullable */
  worstPredictionId: number | null;
}