# lib/api-zod/src/generated/types/simulateAntennaBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Antenna Array Optimizer API
 * OpenAPI spec version: 0.1.0
 */
import type { AntennaConfig } from "./antennaConfig";

export interface SimulateAntennaBody {
  config: AntennaConfig;
  /** Steering angle in degrees */
  theta0Deg?: number;
}