# lib/api-zod/src/generated/types/simulationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Antenna Array Optimizer API
 * OpenAPI spec version: 0.1.0
 */
import type { AntennaConfig } from "./antennaConfig";
import type { PerformanceMetrics } from "./performanceMetrics";
import type { RadiationPattern } from "./radiationPattern";

export interface SimulationResult {
  config: AntennaConfig;
  theta0Deg: number;
  pattern: RadiationPattern;
  metrics: PerformanceMetrics;
}